Fix library namespace Adjust Makefile to generate correct files and directory names
