+++ Makefile
@@ -31,11 +31,11 @@ compile:
 		--out $(LIB_DIR) \
 		--plclname Bio::KBase::$(SERVICE_CAPS)::Client \
 		--jsclname javascript/Client \
-		--pyclname biokbase.$(SERVICE).Client \
+		--pyclname biokbase.$(SERVICE_CAPS).Client \
 		--javasrc src \
 		--java \
-		--pysrvname biokbase.$(SERVICE).Server \
-		--pyimplname biokbase.$(SERVICE).Impl;
+		--pysrvname biokbase.$(SERVICE_CAPS).Server \
+		--pyimplname biokbase.$(SERVICE_CAPS).Impl;
 	touch $(LIB_DIR)/biokbase/__init__.py
 	touch $(LIB_DIR)/biokbase/$(SERVICE)/__init__.py
